Move sql schema and query files to db folder
