+++ rolling-shutter/collator/query_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/jackc/pgx/v4"
 	"gotest.tools/assert"
 
-	"github.com/shutter-network/rolling-shutter/rolling-shutter/collator/cltrdb"
+	"github.com/shutter-network/rolling-shutter/rolling-shutter/db/cltrdb"
 	"github.com/shutter-network/rolling-shutter/rolling-shutter/medley/testdb"
 )
