+++ rolling-shutter/db/cltrdb/db.sqlc.gen.go
