+++ rolling-shutter/db/cltrdb/misc.go
