+++ rolling-shutter/db/cltrdb/models.sqlc.gen.go
