+++ rolling-shutter/db/cltrdb/query.sql
