+++ rolling-shutter/db/cltrdb/query.sqlc.gen.go
