+++ rolling-shutter/db/cltrdb/schema.sql
