+++ rolling-shutter/db/commondb/commondb.go
