+++ rolling-shutter/db/commondb/db.sqlc.gen.go
