+++ rolling-shutter/db/commondb/meta-query.sql
