+++ rolling-shutter/db/commondb/meta-schema.sql
