+++ rolling-shutter/db/commondb/models.sqlc.gen.go
