+++ rolling-shutter/db/commondb/observe-query.sql
