+++ rolling-shutter/db/commondb/observe-query.sqlc.gen.go
