+++ rolling-shutter/db/commondb/observe-schema.sql
