+++ rolling-shutter/db/kprdb/db.sqlc.gen.go
