+++ rolling-shutter/db/kprdb/extend.go
