+++ rolling-shutter/db/kprdb/meta-query.sqlc.gen.go
