+++ rolling-shutter/db/kprdb/models.sqlc.gen.go
