+++ rolling-shutter/db/kprdb/observe-query.sqlc.gen.go
