+++ rolling-shutter/db/kprdb/query.sql
