+++ rolling-shutter/db/kprdb/query.sqlc.gen.go
