+++ rolling-shutter/db/kprdb/schema.sql
