+++ rolling-shutter/sqlc.yaml
@@ -1,18 +1,18 @@
 version: 1
 packages:
-  - path: "keyper/kprdb"
+  - path: "db/kprdb"
     name: "kprdb"
     schema:
       [
-        "keyper/kprdb/schema.sql",
-        "commondb/meta-schema.sql",
-        "commondb/observe-schema.sql",
+        "db/kprdb/schema.sql",
+        "db/commondb/meta-schema.sql",
+        "db/commondb/observe-schema.sql",
       ]
     queries:
       [
-        "keyper/kprdb/query.sql",
-        "commondb/meta-query.sql",
-        "commondb/observe-query.sql",
+        "db/kprdb/query.sql",
+        "db/commondb/meta-query.sql",
+        "db/commondb/observe-query.sql",
       ]
     engine: "postgresql"
     sql_package: "pgx/v4"
@@ -20,20 +20,20 @@ packages:
     output_models_file_name: "models.sqlc.gen.go"
     output_files_suffix: "c.gen"
 
-  - path: "collator/cltrdb"
+  - path: "db/cltrdb"
     name: "cltrdb"
-    schema: ["collator/cltrdb/schema.sql", "commondb/meta-schema.sql"]
-    queries: ["collator/cltrdb/query.sql", "commondb/meta-query.sql"]
+    schema: ["db/cltrdb/schema.sql", "db/commondb/meta-schema.sql"]
+    queries: ["db/cltrdb/query.sql", "db/commondb/meta-query.sql"]
     engine: "postgresql"
     sql_package: "pgx/v4"
     output_db_file_name: "db.sqlc.gen.go"
     output_models_file_name: "models.sqlc.gen.go"
     output_files_suffix: "c.gen"
 
-  - path: "commondb"
+  - path: "db/commondb"
     name: "commondb"
-    schema: ["commondb/meta-schema.sql", "commondb/observe-schema.sql"]
-    queries: ["commondb/meta-query.sql", "commondb/observe-query.sql"]
+    schema: ["db/commondb/meta-schema.sql", "db/commondb/observe-schema.sql"]
+    queries: ["db/commondb/meta-query.sql", "db/commondb/observe-query.sql"]
     engine: "postgresql"
     sql_package: "pgx/v4"
     output_db_file_name: "db.sqlc.gen.go"
